fix(1022): Done
It has been around 12hours since I've started this challenge, and it was
an almost nonstop working until I've found a way to get through it. On
the last moment, I've started to take "notes" inside of comments to
especify tell for myself that it was everything alright. The final piece
was an little thing in line 43 - I was returning a and not c as it was
in my mind. It was like "finding an niddle into the barn".