fix(1040): Found the problem on beecrowd→ there was no problem.
The thing is, I first tend to do it here, with a test "input" file, that
I make with information given by the problem on beecrowd platform. To
make the code get all of the information on that file, I usually do a
loop to get every test at once. That said, I didn't erased that for loop
on line 16...68, so the code was running a bunch of times on Beecrowd.
But there I don't need to do it.

I'm happy that it was only that. I was getting a headache trying to find
something else inside the code, but my logic was right.