fix(1045): Code accepted on Platform
The problem was on my `biggestValue` function. When the input sends the
biggest value in the side_c (third position), my function wasn't
returning it, and it was overwritting the bigger to the second value.

Besides that, I've created, the function to return an Integer value, but
the code inputs was on double value. Through fixing that, I was able to
get my code accepted into Beecrowd platform.